+++ unit-tests/api/craig-api.test.js
@@ -83,7 +83,7 @@ describe("craig api", () => {
         },
         {
           name: "craig/versions.tf",
-          data: '##############################################################################\n# Terraform Providers\n##############################################################################\n\nterraform {\n  required_providers {\n    ibm = {\n      source  = "IBM-Cloud/ibm"\n      version = "~>1.56.1"\n    }\n  }\n  required_version = ">=1.3"\n}\n\n##############################################################################\n',
+          data: '##############################################################################\n# Terraform Providers\n##############################################################################\n\nterraform {\n  required_providers {\n    ibm = {\n      source  = "IBM-Cloud/ibm"\n      version = "~>1.61.0"\n    }\n  }\n  required_version = ">=1.5"\n}\n\n##############################################################################\n',
         },
         {
           name: "craig/ssh_keys.tf",
@@ -109,7 +109,7 @@ describe("craig api", () => {
         },
         {
           name: "craig/management_vpc/versions.tf",
-          data: '##############################################################################\n# Terraform Providers\n##############################################################################\n\nterraform {\n  required_providers {\n    ibm = {\n      source  = "IBM-Cloud/ibm"\n      version = "~>1.56.1"\n    }\n  }\n  required_version = ">=1.3"\n}\n\n##############################################################################\n',
+          data: '##############################################################################\n# Terraform Providers\n##############################################################################\n\nterraform {\n  required_providers {\n    ibm = {\n      source  = "IBM-Cloud/ibm"\n      version = "~>1.61.0"\n    }\n  }\n  required_version = ">=1.5"\n}\n\n##############################################################################\n',
         },
         {
           name: "craig/management_vpc/variables.tf",
@@ -139,7 +139,7 @@ describe("craig api", () => {
         },
         {
           name: "craig/workload_vpc/versions.tf",
-          data: '##############################################################################\n# Terraform Providers\n##############################################################################\n\nterraform {\n  required_providers {\n    ibm = {\n      source  = "IBM-Cloud/ibm"\n      version = "~>1.56.1"\n    }\n  }\n  required_version = ">=1.3"\n}\n\n##############################################################################\n',
+          data: '##############################################################################\n# Terraform Providers\n##############################################################################\n\nterraform {\n  required_providers {\n    ibm = {\n      source  = "IBM-Cloud/ibm"\n      version = "~>1.61.0"\n    }\n  }\n  required_version = ">=1.5"\n}\n\n##############################################################################\n',
         },
         {
           name: "craig/workload_vpc/variables.tf",
@@ -293,10 +293,10 @@ describe("craig api", () => {
             "  required_providers {\n" +
             "    ibm = {\n" +
             '      source  = "IBM-Cloud/ibm"\n' +
-            '      version = "~>1.56.1"\n' +
+            '      version = "~>1.61.0"\n' +
             "    }\n" +
             "  }\n" +
-            '  required_version = ">=1.3"\n' +
+            '  required_version = ">=1.5"\n' +
             "}\n" +
             "\n" +
             "##############################################################################\n",